feat: add tranche fee
this commit adds a fee to the tranche BondController contract. It takes
the fee off of the user's deposit, storing the fee in terms of tranche
tokens (all tranches, i.e. a slice of A, B, and Z) in the BondController
contract. All tranche tokens owned by the BondController will be
considered fee tokens. At maturity, these tokens are automatically
redeemed and the resulting underlying collateral is sent to the owner of
the contract.