feat: add deposit limit guardrail
This commit adds a guardrail that limits the amount of collateral which
can be deposited into a given bond. It leaves this value flexible so
that the creator can set it as desired. Note that it is denominated in
terms of the underlying rebasing collateral for simplicity, so does not
exactly always translate to a dollar value. A bond can also be over the
deposit limit, then later be back below it if the collateral suffers
a negative rebase.